add test

+++ application/use_cases/delete_product.py
@@ -1,3 +1,4 @@
+from fastapi import HTTPException
 from domain.repositories.product_repository import ProductRepository
 
 class DeleteProductUseCase:
@@ -5,4 +6,7 @@ class DeleteProductUseCase:
         self.repository = repository
 
     def execute(self, product_id: str):
+        product = self.repository.get(product_id)
+        if not product:
+            raise HTTPException(status_code=404, detail="Producto no encontrado")
         self.repository.delete(product_id)
